requirements.txt
git+https://github.com/huggingface/diffusers
transformers
huggingface_hub
gradio
numpy==1.26.4
accelerate
optimum-quanto